Implement VS Code versions of Rack and Indicator
These implementations are untested, but provide the actual
functionality for the extension. At this point, the tests verify and
solidify the putOnAHat command functions contract with its
collaborators. The value from the tests came mostly from using them as a
tool for thinking about the design (not that that turned out to be
particularly good...).